// src/common/enums/loan.enum.ts
export enum LoanStatus {
  REGISTRATION = "registration",
  CAPTURING = "capturing",
  APPROVAL = "approval",
  DISBURSEMENT = "disbursement",
  ACTIVE = "active",
  COMPLETED = "completed",
  DEFAULTED = "defaulted",
}

export enum PaymentSchedule {
  WEEKLY = "weekly",
  MONTHLY = "monthly",
}

export enum MaritalStatus {
  SINGLE = "single",
  MARRIED = "married",
  DIVORCED = "divorced",
  WIDOWED = "widowed",
}

export enum IdType {
  GHANA_CARD = "Ghana Card",
  VOTERS_ID = "Voters ID",
  PASSPORT = "Passport",
}